requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
psycopg==3.1.18
psycopg2-binary==2.9.9
setuptools==69.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uuid==1.30
djoser==2.2.2